refactor(BA-6466): move AppProxyStatusResponse to common DTO
Move AppProxyStatusResponse from appproxy/coordinator/api/types.py to
common/dto/appproxy_coordinator/v2/status/types.py so that manager can
reference it without directly depending on the appproxy package.
