+++ src/ai/backend/common/dto/appproxy_coordinator/v2/status/types.py
@@ -1,0 +1,8 @@
+from ai.backend.common.types import BackendAISchema
+
+
+class AppProxyStatusResponse(BackendAISchema):
+    """Response from AppProxy /status endpoint."""
+
+    api_version: str
+    advertise_address: str | None = None
